Extract `io_surface` crate and invoke custom callback on frame sample
