+++ crates/capture/Cargo.toml
@@ -10,6 +10,7 @@ identifier = "dev.zed.Capture"
 
 [dependencies]
 gpui = { path = "../gpui" }
+io_surface = { path = "../io_surface" }
 
 block = "0.1"
 cocoa = "0.24"
