+++ crates/capture/build.rs
@@ -24,8 +24,9 @@ fn main() {
         .allowlist_function("CMTimeMake")
         .allowlist_type("SCStreamOutputType")
         .allowlist_type("SCFrameStatus")
+        .allowlist_type("dispatch_queue_t")
         .allowlist_var("SCStreamFrameInfo.*")
-        .allowlist_function("dispatch_queue_create")
+        .allowlist_var("_dispatch_main_q")
         .parse_callbacks(Box::new(bindgen::CargoCallbacks))
         .layout_tests(false)
         .generate()
