+++ crates/gpui/src/platform/mac.rs
@@ -1,5 +1,5 @@
 mod atlas;
-pub mod dispatcher;
+mod dispatcher;
 mod event;
 mod fonts;
 mod geometry;
